+++ src/ctypesgen/ctypedescs.py
@@ -263,7 +263,7 @@ class CtypesFunction(CtypesType):
         super(CtypesFunction, self).visit(visitor)
 
     def py_string(self, ignore_can_be_ctype=None):
-        return "CFUNCTYPE(%s, %s)" % (
+        return "CFUNCTYPE(UNCHECKED(%s), %s)" % (
             self.restype.py_string(),
             ", ".join(a.py_string() for a in self.argtypes),
         )