+++ src/ctypesgen/processor/operations.py
@@ -83,7 +83,7 @@ def fix_conflicting_names(data, opts):
     
     status_message("Looking for conflicting names...")
     
-    our_names = {"_libs", "_libs_info"}
+    our_names = {"_libs", "_libs_info", "UNCHECKED"}
     # probably includes a bit more than we actually use ...
     our_names |= {x for x in dir(ctypes) if not x.startswith("_")}